A set of fixes for three layout issues: 4160, 4241, and 4243. Review by: jlabanca (desk check)
git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@7250 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/user/client/ui/StackLayoutPanel.java
@@ -211,7 +211,7 @@ public class StackLayoutPanel extends Composite implements HasWidgets,
   }
 
   private void animate(int duration) {
-    int top = 0, bottom = 0;
+    double top = 0, bottom = 0;
     int i = 0, visibleIndex = -1;
     for (; i < layoutData.size(); ++i) {
       LayoutData data = layoutData.get(i);
